requirements.txt
playwright>=1.40.0
openai>=1.0.0
playwright-stealth>=1.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
google-generativeai>=0.3.0
httpx>=0.25.0
